Allowed expiring connections where idle time == expiration time (#171)
* Allowed expiring connections where idle time == expiration time

On uvloop, the timer is only accurate to a millisecond. This negatively affects the `test_connection_pool_get_connection_info` test which may fail because the code runs so fast that the check is done on the same millisecond the connections have expired. This should have no practical implications elsewhere.

* Synced the sync interface with the async one

Co-authored-by: Tom Christie <tom@tomchristie.com>