Adjust max streams when remote settings change (#652)
* Adjust max streams when remote settings change

If the remote server sends a settings update for
`max_concurrent_streams`, check and potentially update the max number
of streams we will use for a connection.

* Update sync version of http2

* Add aiter_stream sub to iter_stream

* Add test for remote max streams update

Improve testing for remote max streams update

This new test checks whether a change in the max streams setting is
properly handled. The max streams value is changed twice throughout the
mock stream. The first update increases the value while the second
update decreases it.

---------

Co-authored-by: Tom Christie <tom@tomchristie.com>